refactor: move statistics charts to different component
